refactor(condo): DOMA-7185 refactored names and placement of tasks and utils for "meter" domain
