+++ apps/condo/domains/meter/tasks/deleteReadingsOfDeletedMeter.spec.js
