+++ apps/condo/domains/meter/tasks/index.js
@@ -1,15 +1,11 @@
 const { createTask, createCronTask } = require('@open-condo/keystone/tasks')
 
-const { deleteReadingsOfDeletedMeterWorker } = require('./deleteReadingsOfDeletedMeterWorker')
-const { sendSubmitMeterReadingsPushNotificationsTaskWorker } = require('./sendSubmitMeterReadingsPushNotificationsTaskWorker')
-const { sendVerificationDateReminderTaskWorker } = require('./sendVerificationDateReminderTaskWorker')
-
-const deleteReadingsOfDeletedMeterTask = createTask('deleteReadingsOfDeletedMeter', deleteReadingsOfDeletedMeterWorker)
-const sendSubmitMeterReadingsPushNotificationsTask = createCronTask('sendSubmitMeterReadingsPushNotifications', '0 14 * * *', sendSubmitMeterReadingsPushNotificationsTaskWorker)
-const sendVerificationDateReminderTask = createCronTask('sendVerificationDateReminder', '0 14 * * 1-5', sendVerificationDateReminderTaskWorker)
+const { deleteReadingsOfDeletedMeter } = require('./deleteReadingsOfDeletedMeter')
+const { sendSubmitMeterReadingsPushNotifications } = require('./sendSubmitMeterReadingsPushNotifications')
+const { sendVerificationDateReminder } = require('./sendVerificationDateReminder')
 
 module.exports = {
-    sendVerificationDateReminderTask,
-    sendSubmitMeterReadingsPushNotificationsTask,
-    deleteReadingsOfDeletedMeterTask,
+    deleteReadingsOfDeletedMeterTask: createTask('deleteReadingsOfDeletedMeter', deleteReadingsOfDeletedMeter),
+    sendSubmitMeterReadingsPushNotificationsCronTask: createCronTask('sendSubmitMeterReadingsPushNotifications', '0 14 * * *', sendSubmitMeterReadingsPushNotifications),
+    sendVerificationDateReminderCronTask: createCronTask('sendVerificationDateReminder', '0 14 * * 1-5', sendVerificationDateReminder),
 }
